Bound GFX_GetDisplayRefreshRate to maximum DOS rate (120 Hz)
There is no sense in going higher: it meets the highest minimum cut off
for VRR monitors (48Hz) and allows peak VESA rates of 120 Hz.

+++ include/video.h
@@ -57,7 +57,7 @@ typedef void (*GFX_CallBack_t)( GFX_CallBackFunctions_t function );
 bool GFX_Events();
 
 Bitu GFX_GetBestMode(Bitu flags);
-int GFX_GetDisplayRefreshRate();
+int8_t GFX_GetDisplayRefreshRate();
 Bitu GFX_GetRGB(Bit8u red,Bit8u green,Bit8u blue);
 void GFX_SetShader(const char* src);
 Bitu GFX_SetSize(int width, int height, Bitu flags,
